requirements.txt
Flask
Flask-Cors
gunicorn
psycopg2-binary
pandas
openpyxl
python-dotenv

